Give Sound Control some love using STL implementations
Sound Control parameters are collected and stored into a std::vector of pairs (string,int) where the string is the name of the parameter and the integer represents the value of the respective parameter.
Using this implementation, we read from the user input the number of the parameter, representing the place of the respective pair in the vector and the new value.
After that, we edit the pair and push the result in the syfs file, then write the new parameters on the screen.

Same method will be used for HotPlug implementation.

Signed off by: Erik Birgiser <ariesrik96@gmail.com>