Let's use STL Map for displaying sysfs tunner values
Each sysfs parameter has his own value. We use the value of each parameter as an unique key in the map. The existance of each sysfs parameter is verified only once, at insertion. This way, we don't have to verify every time ShowSysfsTunner() is called, in fact, we don't need this function anymore as we only need to print the map using a simple iterator. After user input, we also verify if the key exists using the count() method simplifying things and avoiding SIGSEGV